requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
adversarial-robustness-toolbox=1.18.1=pyhd8ed1ab_0
appdirs=1.4.4=pyhd3eb1b0_0
blas=1.0=mkl
brotli=1.0.9=h5eee18b_8
brotli-bin=1.0.9=h5eee18b_8
brotli-python=1.0.9=py38h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.7.4=py38h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py38h06a4308_0
contourpy=1.0.5=py38hdb19cb5_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cuda-version=12.5=3
cycler=0.11.0=pyhd3eb1b0_0
docker-pycreds=0.4.0=pyhd3eb1b0_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py38h06a4308_0
fonttools=4.51.0=py38h5eee18b_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.3.1=py38h06a4308_0
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.43=py38h06a4308_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
huggingface_hub=0.23.1=py38h06a4308_0
idna=3.7=py38h06a4308_0
importlib-metadata=7.0.1=py38h06a4308_0
importlib_metadata=7.0.1=hd3eb1b0_0
importlib_resources=6.1.1=py38h06a4308_1
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.4=py38h06a4308_0
joblib=1.4.2=py38h06a4308_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.4=py38h6a678d5_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libabseil=20240116.2=cxx17_h6a678d5_0
libblas=3.9.0=1_h86c2bf4_netlib
libbrotlicommon=1.0.9=h5eee18b_8
libbrotlidec=1.0.9=h5eee18b_8
libbrotlienc=1.0.9=h5eee18b_8
libcblas=3.9.0=5_h92ddd45_netlib
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.10.1.7=0
libcurand=10.3.6.82=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libllvm14=14.0.6=hdb19cb5_3
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libprotobuf=4.25.3=he621ea3_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp-base=1.3.2=h5eee18b_0
llvm-openmp=14.0.6=h9e868ea_0
llvmlite=0.41.0=py38he621ea3_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py38h5eee18b_0
matplotlib-base=3.7.2=py38h1128e8f_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py38h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py38h06a4308_0
numba=0.58.1=py38h6a678d5_0
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h9ca470c_1
openssl=3.0.14=h5eee18b_0
packaging=24.1=py38h06a4308_0
pathtools=0.1.2=pyhd3eb1b0_1
pillow=10.3.0=py38h5eee18b_0
pip=24.0=py38h06a4308_0
platformdirs=3.10.0=py38h06a4308_0
pooch=1.7.0=py38h06a4308_0
protobuf=4.25.3=py38h12ddb61_0
psutil=5.9.0=py38h5eee18b_0
pyparsing=3.0.9=py38h06a4308_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.19=h955ad1f_0
python-dateutil=2.9.0post0=py38h06a4308_2
python_abi=3.8=2_cp38
pytorch=2.3.1=py3.8_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pyyaml=6.0.1=py38h5eee18b_0
readline=8.2=h5eee18b_0
requests=2.32.2=py38h06a4308_0
safetensors=0.4.2=py38ha89cbab_1
scikit-learn=0.24.2=py38hacb3eff_1
scipy=1.10.1=py38hf6e8229_1
sentry-sdk=1.9.0=py38h06a4308_0
setproctitle=1.2.2=py38h27cfd23_1004
setuptools=69.5.1=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smmap=4.0.0=pyhd3eb1b0_0
sqlite=3.45.3=h5eee18b_0
sympy=1.12=py38h06a4308_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.5.0=py38h2f386ee_0
timm=1.0.7=pyhd8ed1ab_0
tk=8.6.14=h39e8969_0
torchaudio=2.3.1=py38_cu118
torchtriton=2.3.1=py38
torchvision=0.18.1=py38_cu118
tqdm=4.66.4=py38h2f386ee_0
typing-extensions=4.11.0=py38h06a4308_0
typing_extensions=4.11.0=py38h06a4308_0
unicodedata2=15.1.0=py38h5eee18b_0
urllib3=2.2.2=py38h06a4308_0
wandb=0.16.6=pyhd8ed1ab_0
wheel=0.43.0=py38h06a4308_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zipp=3.17.0=py38h06a4308_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
